refactor(dedicated.server): create a mrtg tile
ref: MANAGER-12284
