+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/servers/server/translations/Messages_fr_FR.json
@@ -52,32 +52,6 @@
   "server_tab_IP_rule_status_CREATION_PENDING": "En création",
   "server_tab_IP_rule_status_REMOVAL_PENDING": "En suppression",
   "server_tab_IP_rule_status_OK": "Activée",
-  "server_tab_STATS": "État du serveur",
-  "server_tab_STATS_loading_fail": "Une erreur est survenue lors du chargement des statistiques",
-  "server_tab_stats_network": "Réseau :",
-  "server_tab_stats_network_public": "Interface réseau publique ({{t0}})",
-  "server_tab_stats_network_private": "vRack ({{t0}})",
-  "server_tab_stats_network_no_vrack": "vRack (non configuré sur ce serveur)",
-  "server_tab_stats_type": "Type :",
-  "server_tab_stats_type_ERRORS": "Erreurs",
-  "server_tab_stats_type_PACKETS": "Paquets",
-  "server_tab_stats_type_TRAFFIC": "Trafic",
-  "server_tab_stats_period": "Periode :",
-  "server_tab_stats_period_HOURLY": "dernière heure",
-  "server_tab_stats_period_DAILY": "24 dernières heures",
-  "server_tab_stats_period_WEEKLY": "7 derniers jours",
-  "server_tab_stats_period_MONTHLY": "30 derniers jours",
-  "server_tab_stats_period_YEARLY": "12 derniers mois",
-  "server_tab_STATS_legend_upload": "Upload",
-  "server_tab_STATS_legend_download": "Download",
-  "server_tab_STATS_legend_y_ERRORS": "Nombre d'erreurs",
-  "server_tab_STATS_legend_y_PACKETS": "Nombre de paquets",
-  "server_tab_STATS_legend_y_TRAFFIC": "Niveau de traffic",
-  "server_tab_STATS_legend_no_data": "Pas de données",
-  "server_tab_stats_no_chart": "Impossible d'afficher les données",
-  "server_tab_stats_title": "Graphiques d'utilisation réseaux - MRTG",
-  "server_tab_stats_title_info": "Multi Router Traffic Grapher - logiciel permettant de créer des graphiques sur le trafic réseau",
-  "server_tab_stats_legend_mrtg": "MRTG",
   "server_configuration_state_info_HACKED": "Votre serveur a été hacké et par conséquent placé en rescue ftp. Merci de vérifier votre configuration avant de basculer votre serveur sur disque.",
   "server_configuration_state_info_HACKED_BLOCKED": "Votre serveur a été hacké. Contactez le support pour connaitre la marche à suivre.",
   "server_configuration_renew_service": "Renouveler",
@@ -827,7 +801,6 @@
   "server_dashboard_change_prices": "Votre serveur change de prix,",
   "server_dashboard_change_price_click_here": "cliquez ici",
   "server_dashboard_change_price_more": "pour en savoir plus.",
-
   "server_close_service_form_label": "Sélectionner l'action à effectuer",
   "server_close_service_form_option_terminate": "Supprimer le service immédiatement",
   "server_close_service_form_option_delete_at_expiration": "Supprimer le service à la date d'expiration",
@@ -873,7 +846,6 @@
   "server_datacenter_GSW": "Clichy",
   "server_datacenter_YNM": "Mumbai",
   "server_usage_chart_yaxis_label": "Utilisation (en %)",
-  "server_usage_chart_xaxis_label": "Temps",
   "server_duration_upto": "Jusqu'au {{t0}}",
   "server_error_ola_activation": "Une erreur est survenue lors de l'activation de votre service : {{error}}",
   "server_error_ola_terminate": "Une erreur est survenue lors de la résiliation de votre service : {{error}}",
