+++ packages/manager/modules/bm-server-components/package.json
@@ -12,8 +12,10 @@
   "author": "OVH SAS",
   "main": "./src/index.js",
   "dependencies": {
+    "@ovh-ux/manager-config": "^7.1.1",
     "bootstrap4": "twbs/bootstrap#v4.0.0",
-    "lodash": "^4.17.11"
+    "lodash": "^4.17.11",
+    "moment": "^2.24.0"
   },
   "peerDependencies": {
     "@ovh-ux/manager-core": "^12.0.0 || ^13.0.0",
