+++ packages/manager/modules/bm-server-components/src/index.js
@@ -7,6 +7,7 @@ export { default as serverOrderKvm } from './ipmi/order-kvm';
 export { default as serverNetboot } from './netboot';
 export { default as serverNetwork } from './network-tile';
 export { default as serverConsumptionTile } from './consumption-tile';
+export { default as serverMrtgTile } from './mrtg-tile';
 export { default as serverGeneralInformation } from './general-information';
 export { default as serverTechnicalDetails } from './technical-details';
 export { default as serverServiceStatus } from './service-status';
