+++ packages/manager/modules/bm-server-components/src/mrtg-tile/component.js
@@ -1,0 +1,10 @@
+import controller from './controller';
+import template from './template.html';
+
+export default {
+  bindings: {
+    serverService: '<',
+  },
+  controller,
+  template,
+};
