+++ packages/manager/modules/bm-server-components/src/mrtg-tile/translations/Messages_fr_FR.json
@@ -1,0 +1,29 @@
+{
+  "server_mrtg_loading_fail": "Une erreur est survenue lors du chargement des statistiques",
+  "server_mrtg_network": "Réseau :",
+  "server_mrtg_network_public": "Interface réseau publique ({{t0}})",
+  "server_mrtg_network_private": "vRack ({{t0}})",
+  "server_mrtg_network_no_vrack": "vRack (non configuré sur ce serveur)",
+  "server_mrtg_type": "Type :",
+  "server_mrtg_type_ERRORS": "Erreurs",
+  "server_mrtg_type_PACKETS": "Paquets",
+  "server_mrtg_type_TRAFFIC": "Trafic",
+  "server_mrtg_period": "Periode :",
+  "server_mrtg_period_HOURLY": "dernière heure",
+  "server_mrtg_period_DAILY": "24 dernières heures",
+  "server_mrtg_period_WEEKLY": "7 derniers jours",
+  "server_mrtg_period_MONTHLY": "30 derniers jours",
+  "server_mrtg_period_YEARLY": "12 derniers mois",
+  "server_mrtg_legend_upload": "Upload",
+  "server_mrtg_legend_download": "Download",
+  "server_mrtg_legend_y_ERRORS": "Nombre d'erreurs",
+  "server_mrtg_legend_y_PACKETS": "Nombre de paquets",
+  "server_mrtg_legend_y_TRAFFIC": "Niveau de traffic",
+  "server_mrtg_legend_no_data": "Pas de données",
+  "server_mrtg_no_chart": "Impossible d'afficher les données",
+  "server_mrtg_title": "Graphiques d'utilisation réseaux - MRTG",
+  "server_mrtg_title_info": "Multi Router Traffic Grapher - logiciel permettant de créer des graphiques sur le trafic réseau",
+  "server_mrtg_legend_mrtg": "MRTG",
+  "server_mrtg_unit_KB": "ko / s",
+  "server_mrtg_chart_xaxis_label": "Temps"
+}
